requirements.txt
mysqlclient==2.2.4
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==4.2
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
redis==5.1.0
rq==1.16.2
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
wheel==0.44.0
